squash pull elghoto-version into dev branch (#13)
* Allow setting the source url (for WEB uploads) during metadata choice

Prefix the choice or the url with `*`

* Fix crash when not selecting a source

* Fix error message for 16bit files

* Use `SIMULTANEOUS_THREADS` to specify the number of concurrent threads to use

for operations like downconverting, transcoding, spectrals generation, flac checking...

* Put process_files into common

* Sanitize with configured compression level

* Add search string info for metadata

* Avoid blocking for better progress bar output

* Fix mangled spectrals when file renaming was occuring

This was happening when initial files were named 1. xx, 2. xx, up to 10. xx
Lexical sorting was used, making the spectrals ids 1, 10, 2...

If file renaming was occuring, the spectrals uploaded on torrent description
were not matching the actual files.

* Improve cosmetics between steps

* Add some prettier icons for FLAC/itunes/qobuz

* Fix JunoDownload scraping

* Fix (kinda) itunes scraping

Still broken for artists, explicit tag and discnumber

---------

Co-authored-by: redusys <redusys.ops@gmail.com>